:sparkles: add 회원가입 api

+++ src/main/java/community/mingle/api/domain/member/repository/UniversityRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UniversityRepository extends JpaRepository<University, Long> {
+public interface UniversityRepository extends JpaRepository<University, Integer> {
 
 }
